requirements.txt
Flask
psycopg2-binary
Flask-Cors
python-dotenv
googlemaps
scikit-learn==1.7.1
pandas
gunicorn